Move permission code to authorization
